Update dependencies, move OkhttpClientTest
