+++ gradle/libs.versions.toml
@@ -20,7 +20,7 @@ androidx-test-junit = "1.2.1"
 androidx-work = "2.10.2"
 bitfire-cert4android = "b67ba86d31"
 bitfire-dav4jvm = "05fb8ecda6"
-bitfire-synctools = "032fe6f090"
+bitfire-synctools = "dd6e3a2fb3"
 compose-accompanist = "0.37.3"
 compose-bom = "2025.06.01"
 dnsjava = "3.6.3"
